Cast off_t to int64_t when printing (off_t might have different size)

+++ src/gemdos.c
@@ -889,7 +889,7 @@ static void restore_file_handle_info(int i, FILE_HANDLE *handle)
 	{
 		handle->bUsed = false;
 		Log_Printf(LOG_WARN, "GEMDOS '%s' handle %d cannot be restored, seek to saved offset %"PRId64" failed for: %s\n",
-			   handle->szMode, i, offset, handle->szActualName);
+			   handle->szMode, i, (int64_t)offset, handle->szActualName);
 		fclose(fp);
 		return;
 	}
